+++ src/modules/tokens.ts
@@ -1,0 +1,5 @@
+import { InjectionToken } from '@angular/core';
+import { ShareButtonsMeta, ShareButtonsOptions } from '../models/share-buttons.models';
+
+export const OPTIONS = new InjectionToken<ShareButtonsOptions>('OPTIONS');
+export const BUTTONS_META = new InjectionToken<ShareButtonsMeta>('BUTTONS_META');